Could Me Up/requirements.txt
asgiref==3.11.0
Django==6.0.1
pycryptodome==3.23.0
pycryptodome-test-vectors==1.0.22
sqlparse==0.5.5
tqdm==4.67.1
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
